Installed morgan
I believe in a previous exercise I installed morgan outside of the project folder. While testing, I hadn't noticed because it was technically accessibly at the time, but I ran into issues when attempting to host the project on Heroku. Looking at the logs seems to suggest this is one of the problems causing critical errors, so hopefully pushing this here and onto Heroku will solve things.